- move more functions to table helpers #1785
